requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-cloud-translate==3.8.4
python-dotenv==1.0.0
python-multipart==0.0.6
fastapi-jwt-auth==0.5.0
setuptools
google-api-core==2.11.1
pytest==7.4.0
